error handling

+++ nss.go
@@ -3,6 +3,7 @@ package certinject
 import (
 	"crypto/sha256"
 	"encoding/hex"
+	"fmt"
 	"io/ioutil"
 	"math"
 	"os"
@@ -19,13 +20,13 @@ var certDir = cflag.String(flagGroup, "nsscertdir", "", "Directory to store "+
 var nssDir = cflag.String(flagGroup, "nssdbdir", "", "Directory that "+
 	"contains NSS's cert9.db.  (Required if nss is set.)")
 
-func injectCertNss(derBytes []byte) {
+func injectCertNss(derBytes []byte) error {
 	if certDir.Value() == "" {
-		log.Fatal("Empty nsscertdir configuration.")
+		return fmt.Errorf("Empty nsscertdir configuration.")
 	}
 
 	if nssDir.Value() == "" {
-		log.Fatal("Empty nssdbdir configuration.")
+		return fmt.Errorf("Empty nssdbdir configuration.")
 	}
 
 	fingerprint := sha256.Sum256(derBytes)
@@ -47,33 +48,36 @@ func injectCertNss(derBytes []byte) {
 		if strings.Contains(string(stdoutStderr), "SEC_ERROR_PKCS11_GENERAL_ERROR") {
 			log.Warn("Temporary SEC_ERROR_PKCS11_GENERAL_ERROR injecting certificate to NSS database; retrying in 1ms...")
 			time.Sleep(1 * time.Millisecond)
-			injectCertNss(derBytes)
-		} else {
-			log.Errorf("Error injecting cert to NSS database: %s\n%s", err, stdoutStderr)
+			return injectCertNss(derBytes)
 		}
+		return fmt.Errorf("Error injecting cert to NSS database: %s\n%s", err, stdoutStderr)
 	}
+	return nil
 }
 
-func cleanCertsNss() {
+func cleanCertsNss() error {
 	if certDir.Value() == "" {
-		log.Fatal("Empty nsscertdir configuration.")
+		return fmt.Errorf("Empty nsscertdir configuration.")
 	}
 
 	if nssDir.Value() == "" {
-		log.Fatal("Empty nssdbdir configuration.")
+		return fmt.Errorf("Empty nssdbdir configuration.")
 	}
 
 	certFiles, err := ioutil.ReadDir(certDir.Value() + "/")
 	if err != nil {
-		log.Fatalf("Error enumerating files in cert directory: %s", err)
+		return fmt.Errorf("Error enumerating files in cert directory: %v", err)
 	}
 
 	// for all Namecoin certs in the folder
+	var errors []error
+	var retry = false
 	for _, f := range certFiles {
 		// Check if the cert is expired
 		expired, err := checkCertExpiredNss(f)
 		if err != nil {
-			log.Fatalf("Error checking if NSS cert is expired: %s", err)
+			errors = append(errors, fmt.Errorf("Error checking if NSS cert is expired: %v", err))
+			continue
 		}
 
 		// delete the cert if it's expired
@@ -100,18 +104,32 @@ func cleanCertsNss() {
 			case strings.Contains(string(stdoutStderr), "SEC_ERROR_PKCS11_GENERAL_ERROR"):
 				log.Warn("Temporary SEC_ERROR_PKCS11_GENERAL_ERROR deleting certificate from NSS database; retrying in 1ms...")
 				time.Sleep(1 * time.Millisecond)
-				cleanCertsNss()
+				_, err = cmd.CombinedOutput()
+				if err != nil {
+					errors = append(errors, err)
+					retry = true
+					continue
+				}
 			default:
-				log.Fatalf("Error deleting cert from NSS database: %s\n%s", err, stdoutStderr)
+				log.Fatalf("Error deleting cert from NSS database: %v\n%s", err, stdoutStderr)
 			}
 
-			// Also delete the cert from the filesystem
-			err = os.Remove(certDir.Value() + "/" + filename)
-			if err != nil {
-				log.Fatalf("Error deleting NSS cert from filesystem: %s", err)
+			if !retry {
+
+				// Also delete the cert from the filesystem
+				err = os.Remove(certDir.Value() + "/" + filename)
+				if err != nil {
+					return fmt.Errorf("Error deleting NSS cert from filesystem: %v", err)
+				}
 			}
 		}
 	}
+
+	if len(errors) != 0 {
+		return fmt.Errorf("%v errors found: %v", len(errors), errors)
+	}
+	return nil
+
 }
 
 func checkCertExpiredNss(certFile os.FileInfo) (bool, error) {
